Move Order Command Parsers into order package
