+++ src/main/java/seedu/address/logic/parser/order/OrderFindCommandParser.java
@@ -6,7 +6,7 @@ import static seedu.address.logic.parser.CliSyntax.PREFIX_PHONE;
 
 import java.util.Arrays;
 
-import seedu.address.logic.commands.order.FindOrderCommand;
+import seedu.address.logic.commands.order.OrderFindCommand;
 import seedu.address.logic.commands.order.OrderCommand;
 import seedu.address.logic.parser.ArgumentMultimap;
 import seedu.address.logic.parser.ArgumentTokenizer;
@@ -16,15 +16,15 @@ import seedu.address.model.order.OrderNameContainsKeywordPredicate;
 import seedu.address.model.order.OrderPhoneContainsKeywordPredicate;
 
 /**
- * Parses the given {@code String} of arguments in the context of the FindOrderommand
- * and returns an FindOrderCommand object for execution.
+ * Parses the given {@code String} of arguments in the context of the OrderFindCommand
+ * and returns an OrderFindCommand object for execution.
  * @throws ParseException if the user input does not conform the expected format
  */
-public class FindOrderCommandParser implements Parser<OrderCommand> {
+public class OrderFindCommandParser implements Parser<OrderCommand> {
 
     /**
-     * Parses the given {@code String} of arguments in the context of the FindOrderCommand
-     * and returns an FindOrderCommand object for execution.
+     * Parses the given {@code String} of arguments in the context of the OrderFindCommand
+     * and returns an OrderFindCommand object for execution.
      * @throws ParseException if the user input does not conform the expected format
      */
     @Override
@@ -35,12 +35,12 @@ public class FindOrderCommandParser implements Parser<OrderCommand> {
         if (argMultimap.getValue(PREFIX_NAME).isPresent()) {
             String name = argMultimap.getValue(PREFIX_NAME).get().trim();
             String[] nameKeywords = name.split("\\s+");
-            return new FindOrderCommand(new OrderNameContainsKeywordPredicate(Arrays.asList(nameKeywords)));
+            return new OrderFindCommand(new OrderNameContainsKeywordPredicate(Arrays.asList(nameKeywords)));
         } else if (argMultimap.getValue(PREFIX_PHONE).isPresent()) {
             String phone = argMultimap.getValue(PREFIX_PHONE).get().trim().replaceAll("\\s+", "");
-            return new FindOrderCommand(new OrderPhoneContainsKeywordPredicate(phone));
+            return new OrderFindCommand(new OrderPhoneContainsKeywordPredicate(phone));
         } else {
-            throw new ParseException(String.format(MESSAGE_INVALID_COMMAND_FORMAT, FindOrderCommand.MESSAGE_USAGE));
+            throw new ParseException(String.format(MESSAGE_INVALID_COMMAND_FORMAT, OrderFindCommand.MESSAGE_USAGE));
         }
     }
 }
